# apps/api-gateway/openapi.yaml
openapi: 3.1.0
info:
  title: AGRORED API Gateway
  version: 0.1.0
  description: Punto de entrada del ecosistema AGRORED para enrutar peticiones a los microservicios.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/producers/register:
    post:
      summary: Registrar productor
      tags: [Producers]
      responses:
        "201":
          description: Productor registrado
        "400":
          description: Payload invalido
        "409":
          description: Productor ya existe
  /api/v1/producers:
    get:
      summary: Listar productores
      tags: [Producers]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: Lista paginada de productores
  /api/v1/producers/{id}:
    get:
      summary: Obtener productor por ID
      tags: [Producers]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Detalle del productor
        "404":
          description: No encontrado
...
  /api/v1/rescues/register:
    post:
      summary: Registrar rescate alimentario
      tags: [Rescues]
      responses:
        "201":
          description: Rescate registrado
        "400":
          description: Payload invalido
  /api/v1/rescues:
    get:
      summary: Listar rescates
      tags: [Rescues]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
      responses:
        "200":
          description: Lista paginada de rescates
  /api/v1/rescues/{id}:
    get:
      summary: Obtener rescate por ID
      tags: [Rescues]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Detalle del rescate
        "404":
          description: No encontrado
...